backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-multipart
ultralytics



pip uninstall numpy -y
pip install numpy==1.26.4
pip install tensorflow==2.15
pip install fastapi uvicorn pillow